Changed test to watch for collision persistence instead of collision begin. (#10002)
This test was intermittently failing on AR runs, but it's not immediately clear why. It's possible that the OnCollisionBegin signal could be triggered *before* the test starts listening, which would cause the test to fail. By listening for collision persistence, this test should more robustly handle timing issues.
On the off chance that the timeout was too low in some scenarios, I've raised that as well. This seems much more unlikely to be the culprit, but it doesn't hurt to raise it, since the timeout only affects failures, not success.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>