[Terrain] Remove xy terrain world bounds (#10977)
This PR removes the X/Y bounds from the terrain system since they are no longer needed by any underlying system. The terrain system only needs to know the z bounds to keep the height ranges within reasonable limits for numerical precision. Most of the changes involved relying on a new bus to get the height bounds instead of the world bounds.

A getter for the world bounds still exists, but instead of returning a fixed bounds set by the user, it's simply the combination of the bounds of all terrain areas in the system. This can be used to early out in cases where a point is outside any terrain areas.

Other small unrelated fixes:

- There was an issue with the ray tracing terrain mesh where the wrong bounds were being clamped which could lead to accessing memory out of range.
- The Terrain feature processor had some dead code which has now been removed.